requirements.txt
streamlit
fastapi
uvicorn
pydantic
sqlalchemy
requests
python-dotenv
gdown
papermill
google-api-python-client
google-auth
google-auth-oauthlib
